+++ icons/Icon/IconProps.ts
@@ -8,15 +8,16 @@ export interface InternalIconProps
    * Give the icon a description. It will show up in a tooltip after hovering over the icon for 3s.
    * This is a default browser feature and very helpful to users to understand what the icon does.
    */
+
   title?: string;
   /**
    * If the icon is labelled by an independent element, then provide that element's id. This is
    * used to set the aria-labelledby value on the icon svg. If the `title` prop is also provided,
    * then `titleId` will be used to set the id on the `<title>` element.
    */
+
   titleId?: string;
 }
-
 export type IconProps = {
   /** Identifies the icon */
   kind: IconKind;
@@ -24,12 +25,13 @@ export type IconProps = {
    * Specifies the icon size. The actual icon size is determined based on this prop and the
    * current scale.
    */
+
   size: IconSize;
   /** Set the fill color of the icon */
   // TODO: update type to allowed colors
+
   fill?: string;
 } & InternalIconProps;
-
 export type IconSize = 'L' | 'M' | 'S' | 'XS';
 export type Pixels = 32 | 24 | 16 | 12 | 8;
 type IconKind =
@@ -59,13 +61,18 @@ type IconKind =
   | 'checkboxChecked'
   | 'checkboxEmpty'
   | 'checkmark'
+  | 'circleCheckmark'
   | 'circulatingSupply'
   | 'clock'
   | 'close'
   | 'copy'
   | 'document'
   | 'dot'
+  | 'doubeChevronRight'
+  | 'doubleChevronRight'
   | 'exclamationMark'
+  | 'expand'
+  | 'expandAddress'
   | 'externalLink'
   | 'forward'
   | 'gauge'
@@ -84,6 +91,7 @@ type IconKind =
   | 'lock'
   | 'marketCap'
   | 'minus'
+  | 'more'
   | 'newsFeed'
   | 'notVisible'
   | 'paperAirplane'
@@ -99,10 +107,13 @@ type IconKind =
   | 'pulse'
   | 'qrCode'
   | 'questionMark'
+  | 'reCenter'
   | 'recurring'
   | 'safe'
   | 'search'
   | 'sofort'
+  | 'sortDown'
+  | 'sortUp'
   | 'soundOff'
   | 'soundOn'
   | 'speaker'