+++ icons/react/IconArrowsVertical24Heavy.tsx
@@ -19,6 +19,7 @@ const SvgIconArrowsVertical24Heavy = ({
       width={24}
       height={24}
       fill="none"
+      xmlns="http://www.w3.org/2000/svg"
       className="cds-icon"
       role={role}
       aria-labelledby={titleId}