+++ icons/react/IconCaretUp16Light.tsx
@@ -19,6 +19,7 @@ const SvgIconCaretUp16Light = ({
       width={16}
       height={16}
       fill="none"
+      xmlns="http://www.w3.org/2000/svg"
       className="cds-icon"
       role={role}
       aria-labelledby={titleId}