+++ icons/react/IconMarketCap12Heavy.tsx
@@ -19,6 +19,7 @@ const SvgIconMarketCap12Heavy = ({
       width={12}
       height={12}
       fill="none"
+      xmlns="http://www.w3.org/2000/svg"
       className="cds-icon"
       role={role}
       aria-labelledby={titleId}