+++ icons/react/IconNotVisible12Heavy.tsx
@@ -19,6 +19,7 @@ const SvgIconNotVisible12Heavy = ({
       width={13}
       height={12}
       fill="none"
+      xmlns="http://www.w3.org/2000/svg"
       className="cds-icon"
       role={role}
       aria-labelledby={titleId}