+++ icons/react/IconNotVisible16Heavy.tsx
@@ -19,6 +19,7 @@ const SvgIconNotVisible16Heavy = ({
       width={17}
       height={16}
       fill="none"
+      xmlns="http://www.w3.org/2000/svg"
       className="cds-icon"
       role={role}
       aria-labelledby={titleId}