+++ icons/react/IconPaypal16Heavy.tsx
@@ -19,6 +19,7 @@ const SvgIconPaypal16Heavy = ({
       width={16}
       height={16}
       fill="none"
+      xmlns="http://www.w3.org/2000/svg"
       className="cds-icon"
       role={role}
       aria-labelledby={titleId}