+++ icons/react/IconStep916Heavy.tsx
@@ -14,6 +14,7 @@ const SvgIconStep916Heavy = ({ fill = '#1652F0', title, titleId, ...props }: Int
       width={16}
       height={16}
       fill="none"
+      xmlns="http://www.w3.org/2000/svg"
       className="cds-icon"
       role={role}
       aria-labelledby={titleId}